Add Heterogenous 1D reactor model
