+++ asali/reactors/het1d_transient.py
@@ -1,91 +1,52 @@
 import numpy as np
 
+from asali.reactors.het1d import Heterogeneous1DReactor
+from asali.utils.input_parser import ResolutionMethod
 
-class TransientHeterogeneous1DReactor:
-    def __init__(self,
-                 gas,
-                 surf,
-                 pressure,
-                 alfa,
-                 energy,
-                 inlet_mass_flow_rate,
-                 gas_diffusion,
-                 inlet_temperature,
-                 inlet_mass_fraction,
-                 length,
-                 inert_specie_index,
-                 inert_coverage_index,
-                 reactor_shape_object,
-                 solid):
+
+class TransientHeterogeneous1DReactor(Heterogeneous1DReactor):
+    def __init__(self, cantera_input_file, gas_phase_name, surface_phase_name):
         """
-        Class representing Heterogeneous 1D reactor model
-        :param gas: Cantera gas phase object
-        :param surf: Cantera surface phase object
-        :param pressure: Reactor pressure
-        :param alfa: Reactor catalytic load
-        :param energy: Enable/Disable energy balance
-        :param inlet_mass_flow_rate: Reactor inlet mass flow rate
-        :param gas_diffusion: Enable/Disable gas diffusion
-        :param inlet_temperature: Inlet temperature
-        :param inlet_mass_fraction: Inlet mass fraction
-        :param length: Reactor length
-        :param inert_specie_index: Inert specie index
-        :param inert_coverage_index: Inert coverage index
-        :param reactor_shape_object: Reactor shape class object
-        :param solid: Solid class object
+        Class representing Steady State pseudoHomogeneous 1D reactor model
+        :param cantera_input_file: Cantera file path
+        :param gas_phase_name: Cantera gas phase name
+        :param surface_phase_name: Cantera interface phase name
         """
-        self.gas = gas
-        self.surf = surf
-        self.pressure = pressure
-        self.alfa = alfa
-        self.energy = energy
-        self.inlet_mass_flow_rate = inlet_mass_flow_rate
-        self.gas_diffusion = gas_diffusion
+        super().__init__(cantera_input_file=cantera_input_file, gas_phase_name=gas_phase_name,
+                         surface_phase_name=surface_phase_name)
 
-        self.inlet_temperature = inlet_temperature
-        self.inlet_mass_fraction = inlet_mass_fraction
-        self.length = length
+        self.solution_parser.resolution_method = ResolutionMethod.TRANSIENT
 
-        self.inert_specie_index = inert_specie_index
-        self.inert_coverage_index = inert_coverage_index
-
-        self.reactor_shape_object = reactor_shape_object
-        self.solid = solid
-
-        self.n_p = self.length.size
         self.n_s = self.gas.n_species
         self.n_surf = self.surf.n_species
         self.n_v = self.n_s + self.n_s + self.n_surf + 1 + 1
-        self.alg = None
 
-    def estimate_mass_transfer_coefficient(self, viscosity, density, diffusivity):
+    def initial_condition(self):
         """
-        Estimate mass transfer coefficient based on ReactorModel
-        :param viscosity: Gas viscosity in [Pas]
-        :param density: Gas density in [kg/m3]
-        :param diffusivity: Gas mixture diffusivity in [m2/s]
-        :return: Mass transfer coefficient
+        Function creating the initial condition of the Transient solution
+        :return: Vector/Matrix representing the initial conditions
         """
 
-        self.reactor_shape_object.length = self.length
-        return self.reactor_shape_object.estimate_mass_transfer_coefficient(self.inlet_mass_flow_rate,
-                                                                            viscosity,
-                                                                            density,
-                                                                            diffusivity)
+        y0_matrix = np.zeros([self.n_p, self.n_v], dtype=np.float64)
 
-    def estimate_heat_transfer_coefficient(self, viscosity, conductivity, specific_heat):
-        """
-        Estimate heat transfer coefficient based on ReactorModel
-        :param viscosity: Gas viscosity in [Pas]
-        :param conductivity: Gas thermal conductivity in [W/m/K]
-        :param specific_heat: Gas specific heat in [J/kg/K]
-        :return: Heat transfer coefficient
-        """
-        self.reactor_shape_object.length = self.length
-        return self.reactor_shape_object.estimate_heat_transfer_coefficient(self.inlet_mass_flow_rate,
-                                                                            viscosity,
-                                                                            conductivity,
-                                                                            specific_heat)
+        y0_matrix[:, :self.n_s] = self.initial_mass_fraction
+        y0_matrix[:, self.n_s:self.n_s + self.n_s] = self.initial_mass_fraction
+        y0_matrix[:, self.n_s + self.n_s:self.n_s + self.n_s + self.n_surf] = self.initial_coverage
+        y0_matrix[:, -2] = self.initial_temperature
+        y0_matrix[:, -1] = self.initial_solid_temperature
+
+        if not self.energy:
+            self.inlet_temperature = self.initial_temperature
+            y0_matrix[:, -1] = self.initial_temperature
+
+        y0_matrix[0, :self.n_s] = self.inlet_mass_fraction
+        y0_matrix[0, -1] = self.inlet_temperature
+
+        if not self.is_mass_flow_rate:
+            self.gas.TPY = self.inlet_temperature, self.pressure, self.inlet_mass_fraction
+            self.inlet_mass_flow_rate = self.inlet_volumetric_flow_rate * self.gas.density
+
+        return y0_matrix.flatten()
 
     def equations(self, t, y):
         """
@@ -116,8 +77,8 @@ class TransientHeterogeneous1DReactor:
         r_from_surface = np.zeros([self.n_p, self.n_s], dtype=np.float64)
         r_surface = np.zeros([self.n_p, self.n_surf], dtype=np.float64)
         gas_mix_diff = np.zeros([self.n_p, self.n_s], dtype=np.float64)
-        q_gas = np.zeros([self.n_p], dtype=np.float64)
-        q_surface = np.zeros([self.n_p], dtype=np.float64)
+        q_from_gas = np.zeros([self.n_p], dtype=np.float64)
+        q_from_surface = np.zeros([self.n_p], dtype=np.float64)
         gas_rho = np.zeros([self.n_p], dtype=np.float64)
         gas_cp = np.zeros([self.n_p], dtype=np.float64)
         gas_k = np.zeros([self.n_p], dtype=np.float64)
@@ -139,18 +100,16 @@ class TransientHeterogeneous1DReactor:
             gas_mix_diff[i, :] = diff_mix
 
             if self.energy:
-                if self.gas.n_reactions > 0:
-                    q_gas[i] = -np.dot(self.gas.net_rates_of_progress, self.gas.delta_enthalpy)
+                q_from_gas[i] = self.get_homogeneous_heat_of_reaction()
 
             self.gas.TPY = Tw[i], self.pressure, omegaw[i, :]
             self.surf.TP = Tw[i], self.pressure
             self.surf.coverages = z[i, :]
-            r_from_surface[i, :] = self.surf.net_production_rates[:self.n_s] * self.gas.molecular_weights
-            r_surface[i, :] = self.surf.net_production_rates[-self.n_surf:]
+            r_from_surface[i, :] = self.get_heterogeneous_gas_species_reaction_rates() * self.gas.molecular_weights
+            r_surface[i, :] = self.get_surface_species_reaction_rates()
 
             if self.energy:
-                if self.surf.n_reactions > 0:
-                    q_surface[i] = -np.dot(self.surf.net_rates_of_progress, self.surf.delta_enthalpy)
+                q_from_surface[i] = self.get_heterogeneous_heat_of_reaction()
 
         k_mat = self.estimate_mass_transfer_coefficient(gas_mu, gas_rho, gas_mix_diff)
         k_heat = self.estimate_heat_transfer_coefficient(gas_mu, gas_k, gas_cp)
@@ -186,7 +145,7 @@ class TransientHeterogeneous1DReactor:
             else:
                 d1st_Tb_outlet = (Tb[-1] - Tb[-2]) / (self.length[-1] - self.length[-2])
                 dTb[-1] = -(self.inlet_mass_flow_rate / (area * gas_rho[-1])) * d1st_Tb_outlet
-                dTb[-1] = dTb[-1] + q_gas[-1] / (gas_rho[-1] * gas_cp[-1])
+                dTb[-1] = dTb[-1] + q_from_gas[-1] / (gas_rho[-1] * gas_cp[-1])
                 dTb[-1] = dTb[-1] - delta_T[-1] / (void_fraction * gas_rho[-1] * gas_cp[-1])
 
             dTw[-1] = Tw[-1] - Tw[-2]
@@ -224,7 +183,7 @@ class TransientHeterogeneous1DReactor:
             # Equations of BULK energy
             d1st_Tb_backward = (Tb[1:-1] - Tb[:-2]) / d1st_length_backward
             dTb[1:-1] = -(self.inlet_mass_flow_rate / (area * gas_rho[1:-1])) * d1st_Tb_backward
-            dTb[1:-1] = dTb[1:-1] + q_gas[1:-1] / (gas_rho[1:-1] * gas_cp[1:-1])
+            dTb[1:-1] = dTb[1:-1] + q_from_gas[1:-1] / (gas_rho[1:-1] * gas_cp[1:-1])
             dTb[1:-1] = dTb[1:-1] - delta_T[1:-1] / (void_fraction * gas_rho[1:-1] * gas_cp[1:-1])
 
             if self.gas_diffusion:
@@ -239,7 +198,7 @@ class TransientHeterogeneous1DReactor:
             d1st_Tw_backward = (Tw[1:-1] - Tw[:-2]) / d1st_length_backward
 
             dTw[1:-1] = (solid_k / (solid_cp * solid_rho)) * (d1st_Tw_forward - d1st_Tw_backward) / d2nd_length
-            dTw[1:-1] = dTw[1:-1] + self.alfa * q_surface[1:-1] / (solid_cp * solid_rho * (1 - void_fraction))
+            dTw[1:-1] = dTw[1:-1] + self.alfa * q_from_surface[1:-1] / (solid_cp * solid_rho * (1 - void_fraction))
             dTw[1:-1] = dTw[1:-1] + delta_T[1:-1] / (solid_cp * solid_rho * (1 - void_fraction))
 
         dy_matrix = np.zeros(shape=y_matrix.shape, dtype=np.float64)
@@ -319,19 +278,26 @@ class TransientHeterogeneous1DReactor:
         res[diff_mask] = res[diff_mask] - dy[diff_mask]
         return res
 
-    def solve(self, numerical_solver, tspan, initial_conditions):
+    def solve(self, tspan, time_ud):
         """
-        Solve Transient equations
-        :param numerical_solver: Numerical solver object
-        :param tspan: Integration time
-        :param initial_conditions: Reactor initial conditions
+        Solve model
+        :param tspan: Vector representing the integration time
+        :param time_ud: Time unit dimension
+        :return: Vector/Matrix representing the results
         :return: Time vector,
                  Matrix representing the solution in terms of composition, coverage and temperature as function of time and reactor length
         """
         self.alg = self.algebraic_equations()
-        return numerical_solver.solve_dae(self.ode_equations,
-                                          self.equations,
-                                          self.residuals,
-                                          initial_conditions,
-                                          tspan,
-                                          self.alg)
+        x, y = self.numerical_solver.solve_dae(self.ode_equations,
+                                               self.equations,
+                                               self.residuals,
+                                               self.initial_condition(),
+                                               self.uc.convert_to_seconds(tspan, time_ud),
+                                               self.alg)
+
+        self.solution_parser.x = x
+        self.solution_parser.y = y
+        self.solution_parser.length = self.length
+        self.solution_parser.is_solved = True
+        return y
+
