+++ examples/het1d_transient.py
@@ -1,10 +1,10 @@
 import os
 
-from asali.reactors.het1d import Heterogeneous1DReactor
 from asali.plotters.reactor import ReactorPlotter
+from asali.reactors.het1d_transient import TransientHeterogeneous1DReactor
 
 if __name__ == "__main__":
-    h = Heterogeneous1DReactor(os.path.join('files', 'H2-O2-Rh.yaml'), 'gas', 'Rh_surface')
+    h = TransientHeterogeneous1DReactor(os.path.join('files', 'H2-O2-Rh.yaml'), 'gas', 'Rh_surface')
     h.set_length([0, 0.5, 1.0, 1.5, 2.0], 'm')
     h.set_pressure(5, 'bar')
     h.set_catalytic_load(10, '1/m')
@@ -19,7 +19,6 @@ if __name__ == "__main__":
     h.set_energy(True)
     h.set_gas_diffusion(False)
     h.set_verbosity(False)
-    h.set_resolution_method("TRANSIENT")
     h.set_initial_mass_fraction({'O2': 0.4, 'AR': 0.5, 'H2': 0.1})
     h.set_initial_temperature(250, 'degC')
     h.set_packed_bed_reactor(0.3, 'mm', 1, 'cm', 0.75)
