--- tests/Heterogeneous1DReactor/get_coverage.asali
@@ -1,7 +1,0 @@
-1.251377350270816147e-01 6.693028584627817059e-06 8.439507952194805762e-01 2.100040139054067372e-05 3.088377632533611736e-02
-1.252613163252381812e-01 1.125259338718453915e-05 8.439267455583100430e-01 2.092957347736193436e-05 3.077975594737946546e-02
-1.528493228222504163e-01 1.678162015314764644e-05 8.198325350710120585e-01 2.329947019903902817e-05 2.727806101383975501e-02
-2.013867645283502483e-01 2.288202276751542097e-05 7.759373874954000883e-01 2.685195082824388586e-05 2.262611400233543665e-02
-2.971596965557277148e-01 2.850017605595530447e-05 6.864022662233264516e-01 3.211594712119127571e-05 1.637742109780755337e-02
-7.602462189186046659e-01 3.627419011729659235e-05 2.374445720007990202e-01 2.504505771353948498e-05 2.247889599026306733e-03
-7.627270889227637518e-01 4.010079672085207174e-05 2.351610613297488717e-01 2.296674772588069664e-05 2.048782454632499476e-03
