--- tests/Heterogeneous1DReactor/get_mass_fraction.asali
@@ -1,7 +1,0 @@
-1.000000000000000056e-01 4.000000000000000222e-01 -2.230581609999999990e-18 5.000000000000000000e-01
-9.967532060000000582e-02 3.974233400000000138e-01 2.901338930000000021e-03 5.000000000000000000e-01
-9.921771729999999401e-02 3.937917950000000000e-01 6.990487679999999975e-03 5.000000000000000000e-01
-9.848706660000000612e-02 3.879933429999999905e-01 1.351959049999999947e-02 5.000000000000000000e-01
-9.710488000000000453e-02 3.770242930000000103e-01 2.587082690000000115e-02 5.000000000000000000e-01
-7.299673380000000489e-02 1.857017470000000003e-01 2.413015209999999910e-01 5.000000000000000000e-01
-7.026434060000000259e-02 1.640174410000000138e-01 2.657182200000000050e-01 5.000000000000000000e-01
