--- tests/Heterogeneous1DReactor/get_mole_fraction.asali
@@ -1,7 +1,0 @@
-6.647472599999999510e-01 1.675267799999999863e-01 0.000000000000000000e+00 1.677259700000000020e-01
-6.633047600000000488e-01 1.666274500000000103e-01 2.160630000000000189e-03 1.679071599999999997e-01
-6.612664099999999712e-01 1.653566200000000097e-01 5.213759999999999575e-03 1.681632100000000074e-01
-6.579988799999999527e-01 1.633194500000000049e-01 1.010801000000000061e-02 1.685736499999999916e-01
-6.517737900000000195e-01 1.594383799999999907e-01 1.943222000000000005e-02 1.693556099999999898e-01
-5.330894900000000547e-01 8.544375999999999360e-02 1.972027699999999995e-01 1.842639799999999939e-01
-5.183062899999999473e-01 7.622706999999999400e-02 2.193456900000000098e-01 1.861209500000000072e-01
