--- tests/Heterogeneous1DReactor/get_solid_mass_fraction.asali
@@ -1,7 +1,0 @@
-9.999860400000000482e-02 3.999829309999999860e-01 1.743564450000000132e-05 5.000000000000000000e-01
-9.967394009999999693e-02 3.974064449999999971e-01 2.918568219999999845e-03 5.000000000000000000e-01
-9.921573190000000653e-02 3.937674629999999842e-01 7.015244549999999878e-03 5.000000000000000000e-01
-9.848379910000000514e-02 3.879532090000000211e-01 1.356029760000000001e-02 5.000000000000000000e-01
-9.709839030000000371e-02 3.769442310000000185e-01 2.595166600000000171e-02 5.000000000000000000e-01
-7.298120680000000360e-02 1.854916369999999870e-01 2.415135959999999971e-01 5.000000000000000000e-01
-7.025055060000000295e-02 1.638287079999999896e-01 2.659090980000000104e-01 5.000000000000000000e-01
