--- tests/Heterogeneous1DReactor/get_solid_mole_fraction.asali
@@ -1,7 +1,0 @@
-6.647402740000000199e-01 1.675202090000000033e-01 1.297036870000000075e-05 1.677265460000000041e-01
-6.632978569999999907e-01 1.666209349999999978e-01 2.173467139999999981e-03 1.679077399999999998e-01
-6.612564760000000375e-01 1.653472280000000127e-01 5.232249849999999654e-03 1.681640450000000120e-01
-6.579825040000000236e-01 1.633039139999999945e-01 1.013853099999999924e-02 1.685750510000000035e-01
-6.517411159999999537e-01 1.594071820000000084e-01 1.949326529999999974e-02 1.693584370000000006e-01
-5.329956829999999979e-01 8.535022679999999806e-02 1.973833419999999894e-01 1.842707480000000120e-01
-5.182210860000000530e-01 7.614178369999999307e-02 2.195102509999999896e-01 1.861268789999999951e-01
