--- tests/Heterogeneous1DReactor/get_solid_temperature.asali
@@ -1,7 +1,0 @@
-5.356246896599999445e+02
-5.356246896599999445e+02
-5.533396212699999523e+02
-5.811435042699999940e+02
-6.298301379700000098e+02
-9.136228575799999589e+02
-9.136228575799999589e+02
