--- tests/Heterogeneous1DReactor/get_temperature.asali
@@ -1,7 +1,0 @@
-5.231499999999999773e+02
-5.355450759599999628e+02
-5.532269317799999726e+02
-5.809687284599999657e+02
-6.295297417499999710e+02
-9.134131216499999937e+02
-9.134131216499999937e+02
