+++ tests/SteadyStateHeterogeneous1DReactor/get_coverage.asali
@@ -1,0 +1,7 @@
+1.148865889999999973e-01 6.474040840000000104e-06 8.527037200000000539e-01 1.999141090000000157e-05 3.238322600000000107e-02
+1.149361200000000027e-01 8.545636640000000745e-06 8.526988260000000208e-01 1.996255739999999973e-05 3.233654510000000259e-02
+1.236351670000000041e-01 1.087524200000000039e-05 8.453189930000000185e-01 2.078011840000000015e-05 3.101418460000000127e-02
+1.342325020000000035e-01 1.330838819999999989e-05 8.361850320000000503e-01 2.173227539999999998e-05 2.954742559999999976e-02
+1.475644859999999947e-01 1.584444300000000127e-05 8.245084569999999724e-01 2.286839160000000116e-05 2.788834449999999868e-02
+9.598750390000000410e-01 1.140221770000000096e-06 3.882302239999999710e-02 3.301093489999999675e-05 1.267787449999999974e-03
+9.600364460000000433e-01 1.243590260000000077e-06 3.864042279999999746e-02 3.363892430000000125e-05 1.288248859999999900e-03
