+++ tests/SteadyStateHeterogeneous1DReactor/get_results.asali
@@ -1,0 +1,7 @@
+1.000000000000000056e-01 4.000000000000000222e-01 -2.179818029999999984e-20 5.000000000000000000e-01 9.999940770000000090e-02 3.999927580000000038e-01 7.397787259999999661e-06 5.000004369999999643e-01 1.148865889999999973e-01 6.474040840000000104e-06 8.527037200000000539e-01 1.999141090000000157e-05 3.238322600000000107e-02 5.231499999999999773e+02 5.284636619999999994e+02
+9.986019840000000258e-02 3.988905329999999916e-01 1.249268840000000106e-03 5.000000000000000000e-01 9.985960889999999368e-02 3.988833219999999846e-01 1.256629009999999953e-03 5.000004399999999904e-01 1.149361200000000027e-01 8.545636640000000745e-06 8.526988260000000208e-01 1.996255739999999973e-05 3.233654510000000259e-02 5.284296379999999544e+02 5.284636619999999994e+02
+9.970150849999999398e-02 3.976311679999999793e-01 2.667323770000000049e-03 5.000000000000000000e-01 9.970083459999999620e-02 3.976229210000000180e-01 2.675734830000000078e-03 5.000005100000000091e-01 1.236351670000000041e-01 1.087524200000000039e-05 8.453189930000000185e-01 2.078011840000000015e-05 3.101418460000000127e-02 5.344797180000000481e+02 5.345186009999999897e+02
+9.951854540000000526e-02 3.961791709999999966e-01 4.302283760000000061e-03 5.000000000000000000e-01 9.951776210000000400e-02 3.961695789999999939e-01 4.312057230000000164e-03 5.000006019999999607e-01 1.342325020000000035e-01 1.330838819999999989e-05 8.361850320000000503e-01 2.173227539999999998e-05 2.954742559999999976e-02 5.415438649999999825e+02 5.415891209999999774e+02
+9.930315550000000380e-02 3.944698329999999920e-01 6.227011129999999658e-03 5.000000000000000000e-01 9.930222430000000544e-02 3.944584250000000014e-01 6.238623750000000037e-03 5.000007269999999782e-01 1.475644859999999947e-01 1.584444300000000127e-05 8.245084569999999724e-01 2.286839160000000116e-05 2.788834449999999868e-02 5.500276350000000321e+02 5.500817809999999781e+02
+8.842732650000000028e-02 3.081591260000000054e-01 1.034135479999999940e-01 5.000000000000000000e-01 8.841470079999999565e-02 3.079994110000000007e-01 1.035704560000000052e-01 5.000154319999999819e-01 9.598750390000000410e-01 1.140221770000000096e-06 3.882302239999999710e-02 3.301093489999999675e-05 1.267787449999999974e-03 1.741655189999999948e+03 1.742375219999999899e+03
+8.710223270000000406e-02 2.976431649999999873e-01 1.152546019999999977e-01 5.000000000000000000e-01 8.709003690000000075e-02 2.974881030000000037e-01 1.154071390000000058e-01 5.000147200000000236e-01 9.600364460000000433e-01 1.243590260000000077e-06 3.864042279999999746e-02 3.363892430000000125e-05 1.288248859999999900e-03 1.742371740000000045e+03 1.742375219999999899e+03
