+++ tests/SteadyStateHeterogeneous1DReactor/initial_condition.asali
